move svgs to /src
