+++ src/svgs/Delete.tsx
