+++ src/svgs/index.ts
